Generate stable `"use cache"` cache key for promise arguments
When passing promise arguments into a `"use cache"` function, the order
in which they resolve currently influences the cache key, leading to
unnecessary cache misses.

This is because we use React's Flight Reply Client to serialize the
arguments into `FormData`. The form data entries are appended as they
stream in, using an incrementing ID as form field name. When all
arguments are resolved, the form data is serialized into a compact
string that is used as one cache key part.

By sorting the form data fields by their key, before serializing them,
we can ensure a deterministic cache key, regardless of when the
individual promise arguments resolve.